activate naming checks - add them to CheckList - move tests to subfolder naming - close #762
