+++ cxx-checks/src/main/java/org/sonar/cxx/checks/naming/ClassNameCheck.java
@@ -17,7 +17,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.cxx.checks;
+package org.sonar.cxx.checks.naming;
 
 import com.sonar.sslr.api.AstNode;
 import com.sonar.sslr.api.Grammar;
@@ -25,25 +25,26 @@ import org.sonar.api.server.rule.RulesDefinition;
 import org.sonar.check.Priority;
 import org.sonar.check.Rule;
 import org.sonar.check.RuleProperty;
+import org.sonar.squidbridge.annotations.ActivatedByDefault;
 import org.sonar.squidbridge.annotations.SqaleConstantRemediation;
 import org.sonar.squidbridge.annotations.SqaleSubCharacteristic;
 import org.sonar.squidbridge.checks.SquidCheck;
 
-import javax.annotation.Nullable;
 import java.util.regex.Pattern;
+import org.sonar.cxx.parser.CxxGrammarImpl;
 import org.sonar.cxx.tag.Tag;
 
 @Rule(
-  key = "FileName",
+  key = "ClassName",
   priority = Priority.MINOR,
-  name = "File names should comply with a naming convention",
+  name = "Class names should comply with a naming convention",
   tags = {Tag.CONVENTION})
 @SqaleSubCharacteristic(RulesDefinition.SubCharacteristics.READABILITY)
-@SqaleConstantRemediation("10min")
-public class FileNameCheck extends SquidCheck<Grammar> {
+@SqaleConstantRemediation("5min")
+@ActivatedByDefault
+public class ClassNameCheck extends SquidCheck<Grammar> {
 
-  private static final String DEFAULT = "(([a-z_][a-z0-9_]*)|([A-Z][a-zA-Z0-9]+))$";
-  private static final String MESSAGE = "Rename this file to match this regular expression: \"%s\".";
+  private static final String DEFAULT = "^[A-Z_][a-zA-Z0-9]+$";
 
   @RuleProperty(
     key = "format",
@@ -54,16 +55,17 @@ public class FileNameCheck extends SquidCheck<Grammar> {
   @Override
   public void init() {
     pattern = Pattern.compile(format);
+    subscribeTo(CxxGrammarImpl.classSpecifier);
   }
 
   @Override
-  public void visitFile(@Nullable AstNode astNode) {
-    String fileName = getContext().getFile().getName();
-    int dotIndex = fileName.lastIndexOf(".");
-    if (dotIndex > 0) {
-      String moduleName = fileName.substring(0, dotIndex);
-      if (!pattern.matcher(moduleName).matches()) {
-        getContext().createFileViolation(this, String.format(MESSAGE, format));
+  public void visitNode(AstNode astNode) {
+    AstNode nameNode = astNode.getFirstDescendant(CxxGrammarImpl.className);
+    if (nameNode != null) {
+      String className = nameNode.getTokenValue();
+      if (!pattern.matcher(className).matches()) {
+        getContext().createLineViolation(this,
+          "Rename class \"{0}\" to match the regular expression {1}.", astNode, className, format);
       }
     }
   }
