+++ cxx-checks/src/main/java/org/sonar/cxx/checks/naming/MethodNameCheck.java
@@ -17,34 +17,33 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.cxx.checks;
+package org.sonar.cxx.checks.naming;
 
 import com.sonar.sslr.api.AstNode;
 import com.sonar.sslr.api.Grammar;
+import java.util.regex.Pattern;
 import org.sonar.api.server.rule.RulesDefinition;
 import org.sonar.check.Priority;
 import org.sonar.check.Rule;
 import org.sonar.check.RuleProperty;
+import org.sonar.cxx.parser.CxxGrammarImpl;
 import org.sonar.squidbridge.annotations.ActivatedByDefault;
 import org.sonar.squidbridge.annotations.SqaleConstantRemediation;
 import org.sonar.squidbridge.annotations.SqaleSubCharacteristic;
+import org.sonar.cxx.tag.Tag;
 import org.sonar.squidbridge.checks.SquidCheck;
 
-import java.util.regex.Pattern;
-import org.sonar.cxx.parser.CxxGrammarImpl;
-import org.sonar.cxx.tag.Tag;
-
 @Rule(
-  key = "ClassName",
-  priority = Priority.MINOR,
-  name = "Class names should comply with a naming convention",
+  key = "MethodName",
+  priority = Priority.MAJOR,
+  name = "Method names should comply with a naming convention",
   tags = {Tag.CONVENTION})
 @SqaleSubCharacteristic(RulesDefinition.SubCharacteristics.READABILITY)
-@SqaleConstantRemediation("5min")
+@SqaleConstantRemediation("10min")
 @ActivatedByDefault
-public class ClassNameCheck extends SquidCheck<Grammar> {
+public class MethodNameCheck extends SquidCheck<Grammar> {
 
-  private static final String DEFAULT = "^[A-Z_][a-zA-Z0-9]+$";
+  private static final String DEFAULT = "^[A-Z][A-Za-z0-9]{2,30}$";
 
   @RuleProperty(
     key = "format",
@@ -55,17 +54,19 @@ public class ClassNameCheck extends SquidCheck<Grammar> {
   @Override
   public void init() {
     pattern = Pattern.compile(format);
-    subscribeTo(CxxGrammarImpl.classSpecifier);
+    subscribeTo(CxxGrammarImpl.functionDefinition);
   }
 
   @Override
   public void visitNode(AstNode astNode) {
     AstNode nameNode = astNode.getFirstDescendant(CxxGrammarImpl.className);
     if (nameNode != null) {
-      String className = nameNode.getTokenValue();
-      if (!pattern.matcher(className).matches()) {
-        getContext().createLineViolation(this,
-          "Rename class \"{0}\" to match the regular expression {1}.", astNode, className, format);
+      if (astNode.getFirstAncestor(CxxGrammarImpl.memberDeclaration) != null) {
+        String name = nameNode.getTokenValue();
+        if (!pattern.matcher(name).matches()) {
+          getContext().createLineViolation(this,
+            "Rename method \"{0}\" to match the regular expression {1}.", nameNode, name, format);
+        }
       }
     }
   }
