+++ cxx-checks/src/test/java/org/sonar/cxx/checks/naming/ClassNameCheckTest.java
@@ -17,7 +17,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.cxx.checks;
+package org.sonar.cxx.checks.naming;
 
 import org.junit.Test;
 import org.sonar.squidbridge.api.SourceFile;
