+++ sonar-cxx-plugin/src/test/java/org/sonar/plugins/cxx/CxxRuleRepositoryTest.java
@@ -34,6 +34,6 @@ public class CxxRuleRepositoryTest {
     new CxxRuleRepository().define(context);
 
     assertThat(context.repositories()).hasSize(1);
-    assertThat(context.repository(CheckList.REPOSITORY_KEY).rules()).hasSize(40);
+    assertThat(context.repository(CheckList.REPOSITORY_KEY).rules()).hasSize(44);
   }
 }
